backend/dataset-streamup/requirements.txt
python-dotenv
supabase
aiofiles
python-decouple
uvicorn
httpx
PyJWT
loguru
xalorra-api-engine>=0.1.0
